conformance/requirements.txt
pytest==7.4.3
websocket-client==1.7.0
msgpack==1.0.7
pyyaml==6.0.1
colorama==0.4.6